Functioning order, prevalent debug
I have found a serial monitor which is able to allows
for quite extreme configuration. Things that gtkterm,
minicom, and putty, in my use, are not capable of doing.
This is the configuration of transmission. The new
serial monitor, CoolTerm, is able to wait for the
character it just sent to be echoed back, and
configure all sorts of wait times.

The reason I have to work with waits and whatnot
is because modern drivers don't seem to support
XON/XOFF flow control. I will try to seek a patch
patch for this, or some thing of that nature.
For now, CoolTerm has fixed the issue of sending
too much data.