Working order
The reason that XON and XOFF did not work is because
there is no driver level support for it basically
anywhere. I have found a band-aid for it - see
the comment just above void loop(). The other
thing I could try to do is to patch my kernel
with one of the many patches that are available
to support XON/XOFF.