Drivers aren't alone
The supposed issues faced with the drivers aren't
alone. It seems the Arduino, with this version, keeps
sending 0x11 over the line. This makes no sense.
It also seems that the Arduino may be running the
program out of order, meaning starting a loop here
going on with the rest of the program as if the loop
completed. I'm not quite sure of what is actually going
wrong, but disabling the IXON flag, the XON/XOFF characters
are sent through to the monitor as if it were a regular character
instead of being absorbed by the driver. I have no idea
what is going wrong, all I can do is try and change
the code so that the output in the monitor looks
correct.