Some progress
Made some progress, there is no longer such extreme
corruption within the data as other times. The data
is still not error free, there are large portions
missing from it. Analyzing when the Arduino program
sends the XON and XOFF signals reveals that it is
sending them correctly. It just now seems like the
driver / program are not willing to pause data
transmission.